+++ dogArea/Views/HomeView/HomeViewModel.swift
@@ -8,7 +8,6 @@
 import Foundation
 import SwiftUI
 import Combine
-import UserNotifications
 
 enum QuestMotionEventType: String, Equatable {
     case progress
@@ -114,117 +113,6 @@ struct WeatherShieldDailySummary: Equatable {
     let dayKey: String
     let applyCount: Int
     let lastAppliedAtText: String
-}
-
-private enum QuestReminderApplyResult: Equatable {
-    case enabled
-    case disabled
-    case permissionDenied
-    case requiresPermission
-}
-
-private protocol QuestReminderScheduling {
-    func applyDailyReminder(
-        enabled: Bool,
-        allowAuthorizationPrompt: Bool,
-        hour: Int,
-        minute: Int
-    ) async -> QuestReminderApplyResult
-}
-
-private final class QuestReminderPreferenceStore {
-    private let defaults = UserDefaults.standard
-    private let key = "home.quest.reminder.enabled.v1"
-
-    /// 저장된 퀘스트 리마인드 on/off 상태를 반환합니다.
-    var isEnabled: Bool {
-        defaults.object(forKey: key) as? Bool ?? false
-    }
-
-    /// 퀘스트 리마인드 on/off 상태를 저장합니다.
-    func setEnabled(_ enabled: Bool) {
-        defaults.set(enabled, forKey: key)
-    }
-}
-
-private final class LocalQuestReminderScheduler: QuestReminderScheduling {
-    private let center: UNUserNotificationCenter
-    private let requestId = "home.quest.daily.reminder.v1"
-
-    init(center: UNUserNotificationCenter = .current()) {
-        self.center = center
-    }
-
-    /// 하루 1회 퀘스트 리마인드 알림을 설정하거나 해제합니다.
-    func applyDailyReminder(
-        enabled: Bool,
-        allowAuthorizationPrompt: Bool,
-        hour: Int,
-        minute: Int
-    ) async -> QuestReminderApplyResult {
-        guard enabled else {
-            center.removePendingNotificationRequests(withIdentifiers: [requestId])
-            center.removeDeliveredNotifications(withIdentifiers: [requestId])
-            return .disabled
-        }
-
-        let authorization = await ensureAuthorization(allowPrompt: allowAuthorizationPrompt)
-        switch authorization {
-        case .enabled:
-            break
-        case .permissionDenied:
-            return .permissionDenied
-        case .requiresPermission:
-            return .requiresPermission
-        case .disabled:
-            return .permissionDenied
-        }
-
-        center.removePendingNotificationRequests(withIdentifiers: [requestId])
-
-        let content = UNMutableNotificationContent()
-        content.title = "오늘 산책 퀘스트 확인할 시간이에요"
-        content.body = "홈에서 오늘 미션을 확인하고, 짧게 기록해 진행도를 올려보세요."
-        content.sound = .default
-
-        var dateComponents = DateComponents()
-        dateComponents.hour = hour
-        dateComponents.minute = minute
-
-        let trigger = UNCalendarNotificationTrigger(dateMatching: dateComponents, repeats: true)
-        let request = UNNotificationRequest(identifier: requestId, content: content, trigger: trigger)
-
-        return await withCheckedContinuation { continuation in
-            center.add(request) { error in
-                continuation.resume(returning: error == nil ? .enabled : .permissionDenied)
-            }
-        }
-    }
-
-    /// 알림 권한 상태를 확인하고 필요 시 사용자 권한 요청을 실행합니다.
-    private func ensureAuthorization(allowPrompt: Bool) async -> QuestReminderApplyResult {
-        let settings = await withCheckedContinuation { continuation in
-            center.getNotificationSettings { settings in
-                continuation.resume(returning: settings)
-            }
-        }
-
-        switch settings.authorizationStatus {
-        case .authorized, .provisional, .ephemeral:
-            return .enabled
-        case .denied:
-            return .permissionDenied
-        case .notDetermined:
-            guard allowPrompt else { return .requiresPermission }
-            return await withCheckedContinuation { continuation in
-                center.requestAuthorization(options: [.alert, .sound, .badge]) { granted, _ in
-                    continuation.resume(returning: granted ? .enabled : .permissionDenied)
-                }
-            }
-        @unknown default:
-            return .permissionDenied
-        }
-    }
 }
 
 final class HomeViewModel: ObservableObject {
